Rename `defaultWorkerFactory` to `webWorkerFactory`
