+++ src/tsec.exemptions.json
@@ -18,7 +18,7 @@
 		"vs/workbench/contrib/notebook/browser/view/renderers/webviewPreloads.ts"
 	],
 	"ban-worker-calls": [
-		"vs/base/browser/defaultWorkerFactory.ts",
+		"vs/base/browser/webWorkerFactory.ts",
 		"vs/workbench/services/extensions/browser/webWorkerExtensionHost.ts"
 	],
 	"ban-worker-importscripts": [
