+++ src/vs/workbench/services/timer/browser/timerService.ts
@@ -18,7 +18,7 @@ import { IPaneCompositePartService } from '../../panecomposite/browser/panecompo
 import { ViewContainerLocation } from '../../../common/views.js';
 import { TelemetryTrustedValue } from '../../../../platform/telemetry/common/telemetryUtils.js';
 import { isWeb } from '../../../../base/common/platform.js';
-import { createBlobWorker } from '../../../../base/browser/defaultWorkerFactory.js';
+import { createBlobWorker } from '../../../../base/browser/webWorkerFactory.js';
 import { Registry } from '../../../../platform/registry/common/platform.js';
 import { ITerminalBackendRegistry, TerminalExtensions } from '../../../../platform/terminal/common/terminal.js';
